new(pkg/modulebuilder): build type own package
